Start på test af PlaneManager. Faketransponder oprettet

+++ AirTrafficMonitoringSystem/Plane/Plane.cs
@@ -9,7 +9,7 @@ namespace AirTrafficMonitoringSystem.Plane
         public int YPosition { get; set; }
         public int Altitude { get; set; }
         public double HorizontalSpeed { get; set; }
-        public int Heading { get; set; }
+        public double Heading { get; set; }
         public DateTime TimeStamp { get; set; }
 
         public static bool operator ==(Plane p1, Plane p2)
